Rename /Storage to /Storages
