+++ Server/Storages/LocalStorage.cs
@@ -1,4 +1,4 @@
-namespace Sunrise.Server.Storage;
+namespace Sunrise.Server.Storages;
 
 public static class LocalStorage
 {
